DisplayPort: change deviceValues array type
Current deviceValues array is int type. However,
on 64-bit platform, the mixer control type of
SNDRV_CTL_ELEM_TYPE_INTEGER has a size of 8.
This would cause memory corruption issue.
Changing the deviceValues array to long type,
so that the memory corruption and further
security issue can be avoided.

+++ device/src/DisplayPort.cpp
@@ -482,7 +482,7 @@ int32_t DisplayPort::setExtDisplayDevice(struct audio_mixer *mixer, int controll
     const char *ctlNamePrefix = "External Display";
     const char *ctlNameSuffix = "Audio Device";
     char mixerCtlName[MIXER_PATH_MAX_LENGTH] = {0};
-    int deviceValues[2] = {-1, -1};
+    long deviceValues[2] = {-1, -1};
 
     ctlIndex = getDisplayPortCtlIndex(controller, stream);
     if (-EINVAL == ctlIndex) {
@@ -510,7 +510,7 @@ int32_t DisplayPort::setExtDisplayDevice(struct audio_mixer *mixer, int controll
         return -EINVAL;
     }
 
-    PAL_DBG(LOG_TAG,"controller/stream: %d/%d", deviceValues[0], deviceValues[1]);
+    PAL_DBG(LOG_TAG,"controller/stream: %ld/%ld", deviceValues[0], deviceValues[1]);
 
     return mixer_ctl_set_array(ctl, deviceValues, ARRAY_SIZE(deviceValues));
 }
